refact: module error
#82
