+++ optimizers/loss.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import math
-from ._util import criterion_entrypoint, is_criterion
+
 
 # https://discuss.pytorch.org/t/is-this-a-correct-implementation-for-focal-loss-in-pytorch/43327/8
 class FocalLoss(nn.Module):
@@ -121,6 +121,21 @@ class F1Loss(nn.Module):
         f1 = 2 * (precision * recall) / (precision + recall + self.epsilon)
         f1 = f1.clamp(min=self.epsilon, max=1 - self.epsilon)
         return 1 - f1.mean()
+
+CRITERION_ENTRYPOINTS = {
+    'cross_entropy': nn.CrossEntropyLoss,
+    'focal': FocalLoss,
+    'label_smoothing': LabelSmoothingLoss,
+    'f1': F1Loss,
+    'bce': nn.BCEWithLogitsLoss,
+}
+
+def criterion_entrypoint(criterion_name):
+    return CRITERION_ENTRYPOINTS[criterion_name]
+
+
+def is_criterion(criterion_name):
+    return criterion_name in CRITERION_ENTRYPOINTS
 
 
 def create_criterion(criterion_name, **kwargs):
